Fix support for lone dash and enhance errors
This commit fixes supporting "lone dashes" on the command line, e.g. to
be used to indicate usage of standard input/output as I/O channels
instead of files on the filesystem.

It also modifies how the program complains when there are "residual
arguments" in intermediate (non-leaf) commands; the printed message
should be less prone to inducing confusion in the reader.